+++ .gitignore
@@ -1,3 +1,9 @@
-notebooks
+Open Algorithms/SummaryWeather.csv
+
 Classic-Methods/tree.dot
-SummaryWeather.csv
+Classic-Methods/cars.csv
+
+notebooks/SummaryWeather.csv
+notebooks/cars.csv
+notebooks/tree.dot
+notebooks/backpropa.png
