+++ README.md
@@ -1,0 +1,9 @@
+## Basic Machine Learning implementations
+
+#### Instructions:
+
+The folder **notebooks** has all files related to the writing of code stored in *Open Algorithms* and *Classic-Methods* folders.
+
+In **Open Algorithms** there is a compact version of code related to maths inherent to different concepts. This is not a formal or extrict rightful aproximation, but it's a mode to superficially understand those concepts. 
+
+In **Classic-Methods** there are several implementatios of libraries.
